requirements.txt
pyrogram==2.0.27
pymongo[srv]==4.1.1
umongo==3.1.0
motor==3.0.0
marshmallow==3.16.0
tgcrypto==1.2.3
python-dotenv=0.20.0